requirements.txt
# Core dependencies
streamlit>=1.30.0
langchain>=0.1.8
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.8.0
pyyaml>=6.0.1

# Document processors
pypdf>=4.0.1
docx2txt>=0.8
unstructured>=0.12.0
markdown>=3.4.1
beautifulsoup4>=4.12.2
pandas>=2.0.0